fix: rounding adjustment in transfer folio

+++ inn/inn_hotels/doctype/inn_folio_transaction/inn_folio_transaction.py
@@ -13,7 +13,7 @@ from frappe.model.document import Document
 
 class InnFolioTransaction(Document):
     def validate(self):
-        amount = int(self.amount)
+        amount = float(self.amount)
         if amount == 0:
         	raise ValueError("Amount cannot be zero")
